detect-bot-service/requirements.txt
botometer==1.6.1
certifi==2022.5.18.1
cffi==1.15.0
charset-normalizer==2.0.12
idna==3.3
Jinja2==3.1.2
MarkupSafe==2.1.1
numpy==1.22.4
oauthlib==3.2.0
pandas==1.4.2
pycparser==2.21
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
requests==2.27.1
requests-oauthlib==1.3.1
rpy2==3.5.1
six==1.16.0
tweepy==3.10.0
tzdata==2022.1
tzlocal==4.2
urllib3==1.26.9
